+++ src/main/java/roomescape/exception/PaymentException.java
@@ -1,6 +1,7 @@
 package roomescape.exception;
 
 import lombok.Getter;
+import roomescape.infrastructure.payment.PaymentErrorCode;
 
 @Getter
 public class PaymentException extends RuntimeException {
